+++ sdk/attestation/attestation/samples-dev/setAttestationPolicy.ts
@@ -39,7 +39,7 @@ import { DefaultAzureCredential } from "@azure/identity";
 
 // Load environment from a .env file if it exists.
 import * as dotenv from "dotenv";
-import { writeBanner, pemFromBase64 } from "./utils/helpers";
+import { writeBanner } from "./utils/helpers";
 import { createRSAKey, createX509Certificate, generateSha256Hash } from "./utils/cryptoUtils";
 
 import { X509 } from "jsrsasign";
@@ -52,7 +52,7 @@ dotenv.config();
 async function setOpenEnclaveAttestationPolicyAadUnsecured() {
   writeBanner("Set OpenEnclave Attestation Policy - Unsecured policy");
 
-  // Use the customer specified attestion URL.
+  // Use the specified attestion URL.
   const endpoint = process.env.ATTESTATION_AAD_URL;
   if (endpoint === undefined) {
     throw new Error("ATTESTATION_AAD_URL must be defined.");
@@ -105,7 +105,7 @@ async function setOpenEnclaveAttestationPolicyAadUnsecured() {
 async function setOpenEnclaveAttestationPolicyAadSecured() {
   writeBanner("Set Open Enclave Attestation Policy - Secured policy");
 
-  // Use the customer specified attestion URL.
+  // Use the specified attestion URL.
   const endpoint = process.env.ATTESTATION_AAD_URL;
   if (endpoint === undefined) {
     throw new Error("ATTESTATION_AAD_URL must be defined.");
@@ -170,7 +170,7 @@ async function setOpenEnclaveAttestationPolicyAadSecured() {
 async function setSgxEnclaveAttestationPolicyIsolatedSecured() {
   writeBanner("Set SGX Enclave Attestation Policy - Secured policy");
 
-  // Use the customer specified attestion URL.
+  // Use the specified attestion URL.
   const endpoint = process.env.ATTESTATION_ISOLATED_URL;
   if (endpoint === undefined) {
     throw new Error("ATTESTATION_ISOLATED_URL must be defined.");
@@ -239,6 +239,24 @@ async function setSgxEnclaveAttestationPolicyIsolatedSecured() {
   );
 
   console.log("Reset attestation policy. Policy status: ", resetPolicyResult.body.policyResolution);
+}
+
+export type PemType = "CERTIFICATE" | "PRIVATE KEY";
+
+/**
+ *
+ * @param base64 - Base64 encoded DER object to encode as PEM.
+ * @param pemType - PEM object type - typically "CERTIFICATE" |
+ */
+export function pemFromBase64(base64: string, pemType: PemType): string {
+  let pem = "-----BEGIN " + pemType + "-----\n";
+  while (base64 !== "") {
+    pem += base64.substr(0, 64) + "\n";
+    base64 = base64.substr(64);
+  }
+  pem += "-----END " + pemType + "-----\n";
+
+  return pem;
 }
 
 export async function main() {