+++ sdk/attestation/attestation/test/public/attestationTests.spec.ts
@@ -208,15 +208,6 @@ describe("[AAD] Attestation Client", function() {
     }
 
     {
-      // If you say you're handing JSON to the service, it should be JSON.
-      await expect(
-        client.attestOpenEnclave(base64url.decodeString(_openEnclaveReport).subarray(0x10), {
-          runTimeJson: stringToBytes("{ xx: abcdefg, }")
-        })
-      ).to.eventually.be.rejectedWith("runTimeJson value cannot be parsed as JSON");
-    }
-
-    {
       const attestationResult = await client.attestOpenEnclave(
         base64url.decodeString(_openEnclaveReport),
         {